mern crud using image with authorization using jsonwebtoken

+++ app/src/Components/MyDeleteComponent.js
@@ -13,36 +13,42 @@ const MyDeleteComponent = ({ itemId, onUpdate }) => {
       return;
     }
 
-    try {
-      const response = await fetch(
-        `http://localhost:3001/itemremove/${itemId}`,
-        {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: token,
-          },
+    
+      if (window.confirm('Are you sure you want to delete this item?')) {
+        try {
+          const response = await fetch(
+            `http://localhost:3001/itemremove/${itemId}`,
+            {
+              method: "DELETE",
+              headers: {
+                "Content-Type": "application/json",
+                Authorization: token,
+              },
+            }
+          );
+    
+          if (response.ok) {
+            setDeleteStatus("Data deleted successfully");
+            onUpdate();
+            navigate("/find");
+          } else if (response.status === 404) {
+            setDeleteStatus("No data found");
+          } else {
+            setDeleteStatus("Failed to delete data");
+          }
+        } catch (error) {
+          console.error("Error:", error.message);
+          setDeleteStatus("Internal Server Error");
         }
-      );
-
-      if (response.ok) {
-        setDeleteStatus("Data deleted successfully");
-        onUpdate();
-        navigate("/find");
-      } else if (response.status === 404) {
-        setDeleteStatus("No data found");
-      } else {
-        setDeleteStatus("Failed to delete data");
       }
-    } catch (error) {
-      console.error("Error:", error.message);
-      setDeleteStatus("Internal Server Error");
-    }
+
+
+   
   };
 
   return (
     <div>
-      <button onClick={handleDelete} className="updatebtn">
+      <button onClick={handleDelete} className="updatebtn btn btn-danger text-nowrap">
         Delete Data
       </button>
       {deleteStatus && <p>{deleteStatus}</p>}
